requirements.txt
PyJWT==1.7.1
Flask==1.1.2
marshmallow==3.6.0
Werkzeug==1.0.1
pytest==5.4.2
SQLAlchemy==1.3.17
pytest-cov==2.8.1
PyMySQL==0.9.3
marshmallow-sqlalchemy==0.23.0
Flask-SQLAlchemy==2.4.1